+++ CHANGES.rst
@@ -4,8 +4,10 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Added test `test_restapi_annex_type_only_for_meeting_managers`, make sure an
+  annex `content_category` that is restricted to `MeetingManagers` using
+  `content_category.only_for_meeting_managers` is rendered the same way.
+  [gbastien]
 
 1.0a1 (2020-01-10)
 ------------------